Move getEngineFactory test
